# docs/api/openAPI.yaml
components:
  schemas:
...
    DataExchangeRequestDto:
      additionalProperties: false
      properties:
        criticality:
          enum:
          - high
          - medium
          - low
          maxItems: 50
          type: string
        desiredEndDateTime:
          format: date-time
          maxItems: 50
          type: string
        desiredStartDateTime:
          format: date-time
          maxItems: 50
          type: string
        notificationId:
          format: uuid
          maxItems: 50
          type: string
        requestId:
          maxItems: 50
          type: string
        requestedTypes:
          items:
            enum:
            - n-tier
            maxItems: 50
            type: string
          maxItems: 50
          type: array
        text:
          maxItems: 50
          type: string
        timestamp:
          format: date-time
          maxItems: 50
          type: string
        uuid:
          format: uuid
          maxItems: 50
          type: string
      required:
      - criticality
      - desiredEndDateTime
      - desiredStartDateTime
      - notificationId
      - requestedTypes
      - text
      type: object
...
    JsonNode:
      type: object
...
  securitySchemes:
    X-API-KEY:
      in: header
      name: X-API-KEY
      type: apiKey
info:
  title: PURIS FOSS Open API
  version: 6.0.0
openapi: 3.1.0
paths:
...
  /data-exchange-request:
    post:
      description: "Creates a new own data exchange request. \n"
      operationId: createDataExchangeRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataExchangeRequestDto'
              additionalProperties: false
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DataExchangeRequestDto'
                additionalProperties: false
          description: Own Data Exchange Request was created.
        '400':
          description: Malformed or invalid request body.
        '409':
          description: Own Data Exchange Request already exists.
        '500':
          description: Internal Server Error.
      summary: Creates a new own data exchange request
      tags:
      - data-exchange-request-controller
...
  /data-exchange-request/request:
    post:
      operationId: postDataExchangeRequest
      parameters:
      - in: header
        name: edc-bpn
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonNode'
              additionalProperties: false
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
      summary: 'This endpoint receives the DataExchangeRequest 1.0.0 requests. This
        endpoint is meant to be accessed by partners via EDC only. '
      tags:
      - data-exchange-request-api-controller
...
security:
- X-API-KEY: []
servers:
- description: Generated server url
  url: http://localhost:8081/catena